auth_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic[email]==2.4.2
python-dotenv==1.0.0
prometheus-client==0.19.0
structlog==23.2.0
tenacity==8.2.3
jinja2==3.1.2
aiosmtplib==2.0.2
fastapi-mail==1.4.1
email-validator==2.1.0.post1